@@ -21,12 +21,12 @@
         "sentry/sdk": "^3.1",
         "slim/psr7": "^1.0",
         "slim/slim": "^4.3",
-        "symfony/cache": "^5.2",
-        "symfony/config": "^5.0",
+        "symfony/cache": "^6.0",
+        "symfony/config": "^6.0",
         "symfony/console": "^5.0",
-        "symfony/mailer": "^5.3",
-        "symfony/translation": "^5.0",
-        "symfony/validator": "^5.0",
+        "symfony/mailer": "^6.0",
+        "symfony/translation": "^6.0",
+        "symfony/validator": "^6.0",
         "twig/twig": "^3.0",
         "webmozart/assert": "^1.6"
     },
@@ -37,7 +37,7 @@
         "friendsofphp/php-cs-fixer": "^3.0",
         "guzzlehttp/guzzle": "^7.2",
         "lctrs/psalm-psr-container-plugin": "^1.1",
-        "overtrue/phplint": "^3.0",
+        "overtrue/phplint": "^4.0",
         "phpunit/phpunit": "^9.2",
         "psalm/plugin-phpunit": "^0.16.0",
         "roave/security-advisories": "dev-latest",
