@@ -8,14 +8,11 @@ use Symfony\Component\Validator\Constraints as Assert;
 
 final class Command
 {
-    /**
-     * @Assert\NotBlank
-     * @Assert\Email
-     */
+    #[Assert\NotBlank]
+    #[Assert\Email]
     public string $email = '';
-    /**
-     * @Assert\NotBlank
-     * @Assert\Length(min=6, allowEmptyString=true)
-     */
+
+    #[Assert\Length(min: 6)]
+    #[Assert\NotBlank]
     public string $password = '';
 }
