@@ -7,18 +7,13 @@ namespace App\Auth\Entity\User;
 use Doctrine\ORM\Mapping as ORM;
 use Webmozart\Assert\Assert;
 
-/**
- * @ORM\Embeddable
- */
+#[ORM\Embeddable]
 final class Network
 {
-    /**
-     * @ORM\Column(type="string", length=16)
-     */
+    #[ORM\Column(type: 'string', length: 16)]
     private string $name;
-    /**
-     * @ORM\Column(type="string", length=16)
-     */
+
+    #[ORM\Column(type: 'string', length: 16)]
     private string $identity;
 
     public function __construct(string $name, string $identity)
