@@ -9,20 +9,18 @@ use Doctrine\ORM\Mapping as ORM;
 use DomainException;
 use Webmozart\Assert\Assert;
 
-/**
- * @ORM\Embeddable
- */
+#[ORM\Embeddable]
 final class Token
 {
     /**
      * @var string
-     * @ORM\Column(type="string", nullable=true)
      */
+    #[ORM\Column(type: 'string', nullable: true)]
     private $value;
     /**
      * @var DateTimeImmutable
-     * @ORM\Column(type="datetime_immutable", nullable=true)
      */
+    #[ORM\Column(type: 'datetime_immutable', nullable: true)]
     private $expires;
 
     public function __construct(string $value, DateTimeImmutable $expires)
