@@ -11,58 +11,46 @@ use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
 use DomainException;
 
-/**
- * @ORM\Entity
- * @ORM\HasLifecycleCallbacks
- * @ORM\Table(name="auth_users")
- */
-class User
+#[ORM\Entity]
+#[ORM\HasLifecycleCallbacks]
+#[ORM\Table(name: 'auth_users')]
+final class User
 {
-    /**
-     * @ORM\Column(type="auth_user_id")
-     * @ORM\Id
-     */
+    #[ORM\Column(type: 'auth_user_id')]
+    #[ORM\Id]
     private Id $id;
-    /**
-     * @ORM\Column(type="datetime_immutable")
-     */
+
+    #[ORM\Column(type: 'datetime_immutable')]
     private DateTimeImmutable $date;
-    /**
-     * @ORM\Column(type="auth_user_email", unique=true)
-     */
+
+    #[ORM\Column(type: 'auth_user_email', unique: true)]
     private Email $email;
-    /**
-     * @ORM\Column(type="string", nullable=true)
-     */
+
+    #[ORM\Column(type: 'string', nullable: true)]
     private ?string $passwordHash = null;
-    /**
-     * @ORM\Column(type="auth_user_status", length=16)
-     */
+
+    #[ORM\Column(type: 'auth_user_status', length: 16)]
     private Status $status;
-    /**
-     * @ORM\Embedded(class="Token")
-     */
+
+    #[ORM\Embedded(class: Token::class)]
     private ?Token $joinConfirmToken = null;
-    /**
-     * @ORM\Embedded(class="Token")
-     */
+
+    #[ORM\Embedded(class: Token::class)]
     private ?Token $passwordResetToken = null;
-    /**
-     * @ORM\Column(type="auth_user_email", nullable=true)
-     */
+
+    #[ORM\Column(type: 'auth_user_email', nullable: true)]
     private ?Email $newEmail = null;
-    /**
-     * @ORM\Embedded(class="Token")
-     */
+
+    #[ORM\Embedded(class: Token::class)]
     private ?Token $newEmailToken = null;
-    /**
-     * @ORM\Column(type="auth_user_role", length=16)
-     */
+
+    #[ORM\Column(type: 'auth_user_role', length: 16)]
     private Role $role;
+
     /**
      * @var Collection<array-key,UserNetwork>
-     * @ORM\OneToMany(targetEntity="UserNetwork", mappedBy="user", cascade={"all"}, orphanRemoval=true)
      */
+    #[ORM\OneToMany(mappedBy: 'user', targetEntity: UserNetwork::class, cascade: ['all'], orphanRemoval: true)]
     private Collection $networks;
 
     private function __construct(Id $id, DateTimeImmutable $date, Email $email, Status $status)
@@ -253,9 +241,7 @@ class User
         return $this->networks->map(static fn (UserNetwork $network) => $network->getNetwork())->toArray();
     }
 
-    /**
-     * @ORM\PostLoad
-     */
+    #[ORM\PostLoad]
     public function checkEmbeds(): void
     {
         if ($this->joinConfirmToken && $this->joinConfirmToken->isEmpty()) {
