@@ -7,27 +7,20 @@ namespace App\Auth\Entity\User;
 use Doctrine\ORM\Mapping as ORM;
 use Ramsey\Uuid\Uuid;
 
-/**
- * @ORM\Entity(readOnly=true)
- * @ORM\Table(name="auth_user_networks", uniqueConstraints={
- *     @ORM\UniqueConstraint(columns={"network_name", "network_identity"})
- * })
- */
-class UserNetwork
+#[ORM\Entity(readOnly: true)]
+#[ORM\Table(name: 'auth_user_networks')]
+#[ORM\UniqueConstraint(columns: ['network_name', 'network_identity'])]
+final class UserNetwork
 {
-    /**
-     * @ORM\Column(type="guid")
-     * @ORM\Id
-     */
+    #[ORM\Column(type: 'guid')]
+    #[ORM\Id]
     private string $id;
-    /**
-     * @ORM\ManyToOne(targetEntity="User", inversedBy="networks")
-     * @ORM\JoinColumn(name="user_id", referencedColumnName="id", nullable=false, onDelete="CASCADE")
-     */
+
+    #[ORM\ManyToOne(targetEntity: User::class, inversedBy: 'networks')]
+    #[ORM\JoinColumn(name: 'user_id', referencedColumnName: 'id', nullable: false, onDelete: 'CASCADE')]
     private User $user;
-    /**
-     * @ORM\Embedded(class="Network")
-     */
+
+    #[ORM\Embedded(class: Network::class)]
     private Network $network;
 
     public function __construct(User $user, Network $network)
