@@ -10,6 +10,7 @@ use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Types\Type;
 use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\EntityManagerInterface;
+use Doctrine\ORM\Mapping\Driver\AttributeDriver;
 use Doctrine\ORM\Mapping\UnderscoreNamingStrategy;
 use Doctrine\ORM\Tools\Setup;
 use Psr\Container\ContainerInterface;
@@ -22,7 +23,7 @@ return [
         /**
          * @psalm-suppress MixedArrayAccess
          * @var array{
-         *     metadata_dirs:array,
+         *     metadata_dirs:string[],
          *     dev_mode:bool,
          *     proxy_dir:string,
          *     cache_dir:?string,
@@ -33,15 +34,15 @@ return [
          */
         $settings = $container->get('config')['doctrine'];
 
-        $config = Setup::createAnnotationMetadataConfiguration(
-            $settings['metadata_dirs'],
+        $config = Setup::createConfiguration(
             $settings['dev_mode'],
             $settings['proxy_dir'],
             $settings['cache_dir'] ?
                 DoctrineProvider::wrap(new FilesystemAdapter('', 0, $settings['cache_dir'])) :
-                DoctrineProvider::wrap(new ArrayAdapter()),
-            false
+                DoctrineProvider::wrap(new ArrayAdapter())
         );
+
+        $config->setMetadataDriverImpl(new AttributeDriver($settings['metadata_dirs']));
 
         $config->setNamingStrategy(new UnderscoreNamingStrategy());
 
